requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anthropic==0.5.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
decorator==4.4.2
distro==1.8.0
elevenlabs==0.2.24
exceptiongroup==1.1.3
executing==2.0.0
ffmpeg==1.4
ffmpeg-downloader==0.2.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.10.0
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2
numba==0.58.1
numpy==1.26.1
openai==0.28.1
openai-whisper==20230918
packaging==23.2
pafy==0.5.5
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.13
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.34.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
websockets==12.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.17.0
